+++ src/types/MergedMethodConfig.ts
@@ -2,7 +2,9 @@ export interface MergedMethodConfig {
   separator: {
     preLog?: string;
     postLog?: string;
+    color?: string;
   };
   showFileName: boolean;
-  showComponentName: boolean;
+  showFunctionName: boolean;
+  addNewLine: boolean;
 }