+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.2.0-beta.0](https://github.com/Raycas96/loggatron/compare/v1.1.1-alpha.1...v1.2.0-beta.0) (2025-11-07)
+
+### Features
+
+- **config:** add addNewLine configuration option ([41a9092](https://github.com/Raycas96/loggatron/commit/41a90921b38fc0619fbc62fb325015ad9c4ea7cf))
+- **separator:** add color configuration support ([1d01b52](https://github.com/Raycas96/loggatron/commit/1d01b52e70dbd16573e0fd92ca4315f47099bf60))
+
+### Bug Fixes
+
+- use method.Config.separator.color directly ([2290ce5](https://github.com/Raycas96/loggatron/commit/2290ce5364acf492bd7a3ee4b248e2361818a459))
+- **config:** add addNewLine to MergedMethodConfig for method overrides ([c8b054e](https://github.com/Raycas96/loggatron/commit/c8b054e08b7913e64d22c75740d367f0beaa03bc))
+- **package:** remove type module for NestJS compatibility ([5d1a6ec](https://github.com/Raycas96/loggatron/commit/5d1a6ecd3fa3d47827a11b4f6a71aa90e001e149))
+- **separator:** always use console.log for separator output ([4fe36e5](https://github.com/Raycas96/loggatron/commit/4fe36e5cf6daedf39851b86cc010df343c4888fa))
+
 ### [1.1.1-alpha.1](https://github.com/Raycas96/loggatron/compare/v1.1.1-alpha.0...v1.1.1-alpha.1) (2025-11-06)
 
 ### [1.1.1-alpha.0](https://github.com/Raycas96/loggatron/compare/v1.1.0...v1.1.1-alpha.0) (2025-11-06)